requirements.txt
colorama==0.4.6
requests==2.32.4
beautifulsoup4==4.13.4
dnspython==2.7.0
lxml==5.4.0
